requirements.txt
openai
Flask
Scrapy
python-dotenv
os
copy